Safari-609.1.11.1/LayoutTests/memory/memory-pressure-simulation.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<script src="../resources/js-test-pre.js"></script>
</head>
<body>
<script>

description("Basic test for the memory pressure simulation mechanism.");

debug("Verify that we're not running under memory pressure from the beginning.");
shouldBe("internals.isUnderMemoryPressure", "false");

debug("Begin simulated memory pressure.");
internals.beginSimulatedMemoryPressure();

shouldBe("internals.isUnderMemoryPressure", "true");

debug("End simulated memory pressure.");
internals.endSimulatedMemoryPressure();

shouldBe("internals.isUnderMemoryPressure", "false");

</script>
<script src="../resources/js-test-post.js"></script>
</body>
</html>
